+++ src/common/message.ts
@@ -1,0 +1,9 @@
+import * as vscode from 'vscode'
+
+export function show(message: string) {
+    vscode.window.showInformationMessage(message)
+}
+
+export function error(message: string) {
+    vscode.window.showErrorMessage(message)
+}