+++ test/files-to-run/run as.spec.ts
@@ -1,0 +1,6 @@
+describe('Run As', () => {
+    it('Run As Mocha', () => {
+        let [a, b] = ['Hello', 'World']
+        console.log(`${a},${b}!`)
+    })
+})