+++ test/files-to-run/run as.ts
@@ -1,0 +1,2 @@
+let [a, b] = ['Hello', 'World']
+console.log(`${a},${b}!`)